WIP TESTS ARE FAILING
allow objects to be loaded and saved without implementing IContextualNode

If they only need [NodeData] fields to be considered, allow them to have an attribute and call LoadFields/SaveFields automatically

Tests fail because in NodeDataListLibrary it tries to Debug.Log and then when that fails it tries to create fatal error dialog.  It can't call Unity methods of course.